Update dependencies, improve sanitizing methods, and enhance documentation
Updated the version of 'de.eldoria.util' dependency to '2.0.1-DEV' in build.gradle.kts for enhanced features and bug fixes. Added a builder method in Settings.java for easy creation of Settings object. Improved the documentation of SizedReport.java to provide more clarity on its methods. Now using the isEmpty method to get whether the list is empty, instead of comparing size to 0 in SizedReport.java for better readability. Extensively updated the README.md file for detailed instructions on plugin and API usage. Also, added some sanitizing code and methods in SanitizerPlugin.java and Configuration.java for effective sanitizing operations. These changes provide better structure and explanation to the code, making it easier to understand and navigate.