Add GitHub Action to publish artifacts to Nexus, update dependencies, and refactor code
Added a GitHub Action workflow to publish artifacts to Nexus on push events. This allows for continuous delivery of the software and simplifies the publishing process. Updated dependencies in build.gradle.kts to newer versions for enhanced features and bug fixes. Refactored the java-version to use version '17' from Temurin distributions to ensure compatibility with the latest software systems. Additionally, updated Strings to make use of code literals in Javadoc comments enhancing readability and understanding. Added GNU Lesser General Public License file to ensure the open source nature of the project and to provide clarity on the terms and conditions of usage. Refactored some of the code in project modules for better organization and readability. These collectively enhance the usability of the software, provide continual updates, and make the codebase easier to navigate and understand.