Add SPDX-License-Identifier and Copyright information in the header of each file
Added SPDX-License-Identifier and Copyright information to the header of all source files. This change was required to specify license information for these files in line with the LGPL-3.0-or-later license the project is under, and to assert copyright for the project maintainer, RainbowDashLabs and all contributors. A few additions to the build.gradle.kts were also made to support the maven-publish plugin.