Add file existence check before sanitizing
Two significant changes were made. In Sanitizer.java, a FileNotFoundException was added to ensure the the file exists before trying to sanitize it. In SanitizeBase.java, an exception was added to warn the end user when they're trying to sanitize a directory instead of individual file. This was necessary to avoid unnecessary errors and provide a more informative feedback for the user.