Refactor schematic sanitization and improved error handling
Refactored sanitization process by fixing handling of 'o' flag in the 'Fix.java'. Created a new naming practice for the sanitized schematics to eliminate any confusion about which files have been processed.

Reorganized the structure by moving the 'WorldEditUtil.java' into a 'util' package for a more organized structure.

Error feedback was improved: more precise messages are added to guide the user in the case when the file to process is a directory instead of a single file. The failure to create a new directory now provides a detailed log message for troubleshooting.

Also handled exceptions arising from directory creation within 'FixBatch.java'. While processing multiple files, if the directory already exists it is now ignored and in case of other errors, a log message is produced to help with debugging. Additionally, a new directory is created for batch sanitized files for clarity.