Add Load class for report loading
This commit introduces a new Load class within the schematicclean.report package. This class provides utility to load reports and implements both AdvancedCommand and ITabExecutor. The purpose of this class is to streamline the process of loading reports in the application, especially when the users want to fetch a specific report file using the report path.