Add path information to SanitizerReport
Added the `path` field to SanaizerReport class. This track ensures we track the file path of the schematic we analysed. Apart from the `path` field, logic for calculating the percentage of blocks removed has also been included, giving more insight into how much of the original schematic was sanitised. Finally, error handling for block removal has been improved by providing a detailed report if it's not empty. The path is now passed through the sanitization process, from `Sanitizer` through `SanitizerExtent` to `ReportBuilder`, allowing the path information to be included in the report.