Add block removal percentage and full report link to sanitizer report
Added two more pieces of information to the shortComponents function in the SanitizerReport class. Firstly, a percentage is now computed if any blocks have been removed during the sanitization process, and an error message is added if this percentage is greater than or equal to 90%. Secondly, a link has been added to the end of the error messages that opens the full report when clicked. The path argument for the report has also been passed from the ReportBuilder's build function to the constructor of the SanitizerReport. These changes provide more detailed information at a glance and easier access to full details when a large amount of content has been modified.