Add GitHub Actions workflows for verification and javadocs
This commit includes two new GitHub Actions workflows: `verify.yml` and `javadocs.yml`. The `verify.yml` workflow is triggered on every push and pull request to verify the state of the codebase by running tests and building the project on both Ubuntu and Windows environments. The `javadocs.yml` workflow, on the other hand, is designed to build and deploy javadocs to the gh-pages branch. It runs only when changes are pushed to the master branch or manually triggered via workflow dispatch, aiming to keep our javadocs up-to-date with the latest codebase on main.