Refactor code to improve readability and maintainability
Code in several classes of the schematicclean report and sanitizing utilities has been modified for improved organization and maintainability. Unnecessary configurations were removed from the Show class, unused imports and methods were taken out in various classes (Text, Colors, RemovedEntityNbtRenderer). This is done to simplify the codebase and make it easier for future development and debugging.