Add entity dumping to SanitizerPlugin
This update incorporates the ability to dump entities to the 'SanitizerPlugin.java' within the 'Eldoria' package. New imports were added to facilitate this, while the 'onPostStart' method now includes a try-catch block that writes the identities of all entities to 'entity_types.txt.' If this process fails, it logs a warning indicating that it was unable to dump the entities.