Update test case assertions for file list size
The test cases 'completeEmptyFile' and 'completeEmptyFileSlash', in the CompletionTest class of the SchematicSanitizer plugin, have been updated. Change was made to assertions that checked the size of the list of strings retrieved after running 'Completion.completeFiles'. The size has been revised from 3 to 4 to reflect additional files being included in the testing.