Add Hangar publishing to the plugin build
In this commit, the Hangar plugin has been added to the build script for enabling the publication of the plugin on the Hangar platform. We have specifically registered a "plugin" publication, using the project version as its version and a namespace of "Eldoria" and "SchematicSanitizer". The Hangar API key and channel are read from environment variables and a platform of "PAPER" has been registered. Dependencies have also been defined. A task for the shadowJar has been updated with archive base name and version set.